requirements.txt
requests
flask
python-dotenv
pandas
tabulate
beautifulsoup4
anthropic
